add int64_t to cell types

+++ xlcpp-pimpl.h
@@ -135,7 +135,7 @@ public:
     const style* sty;
 
     unsigned int num;
-    std::variant<int, shared_string, double, date, time, datetime> val;
+    std::variant<int64_t, shared_string, double, date, time, datetime> val;
 };
 
 };
